// colaboradoresAngular/src/app/shared/interfaces/responseSkills.ts
export class ReponseSkills {
  'microservicio': string;
  'endpoint': string;
  'url': string;
  'metodo': string;
  'identificacion': string;
  'codigoEstatusHttp': string;
  'estatusHttp': string;
  'codigoEstatus': number;
  'descripcion': string;
  'tiempoEjecucion': string;
  'memoriaConsumida': string;
  'resultados': string;
}